src/backend/requirements.txt
# ── Web 框架 ─────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9          # form data / file upload

# ── 数据库 ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0                  # async PostgreSQL driver
psycopg2-binary>=2.9.9           # Alembic sync 操作用

# ── 任务队列 ────────────────────────────────────────────────────────────────
celery>=5.3.0
celery[redis]>=5.3.0
redis>=5.0.0

# ── 配置 & 校验 ──────────────────────────────────────────────────────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# ── 认证 & 加密 ──────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0  # JWT
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1,<5.0.0              # 避免 bcrypt 5 与 passlib 1.7.4 兼容问题
cryptography>=42.0.0              # AES-256 Cookie 加密

# ── HTTP 客户端（AIGC / 百家号 API 调用）────────────────────────────────────
httpx>=0.26.0
aiohttp>=3.9.0                   # SSE 流式读取

# ── 调度 ────────────────────────────────────────────────────────────────────
croniter>=2.0.0                  # Cron 表达式解析与下次触发时间计算
pytz>=2023.3                     # 时区处理（调度、每日限额计算）

# ── 测试 ────────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
